requirements.txt
dj-database-url==0.5.0
Django==3.1
django-widget-tweaks==1.4.8
et-xmlfile==1.0.1
gunicorn==20.0.4
jdcal==1.4.1
openpyxl==3.0.3
psycopg2==2.8.5
psycopg2-binary==2.8.5
python-dotenv==0.13.0
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.1.0
django-bootstrap-modal-forms==2.0.0
